# server/models/operations.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from server.models.base_model_ import Model
from server.models.operation import Operation  # noqa: F401,E501
from server import util


class Operations(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self):  # noqa: E501
        """Operations - a model defined in Swagger

        """
        self.swagger_types = {
        }

        self.attribute_map = {
        }

    @classmethod
    def from_dict(cls, dikt) -> 'Operations':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Operations of this Operations.  # noqa: E501
        :rtype: Operations
        """
        return util.deserialize_model(dikt, cls)