# server/models/status.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from server.models.base_model_ import Model
from server.models.point import Point  # noqa: F401,E501
from server import util


class Status(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, host: str=None, version: str=None, checked_out: bool=None, checked_out_since: datetime=None, movements_since_checkout: int=None, position: Point=None):  # noqa: E501
        """Status - a model defined in Swagger

        :param host: The host of this Status.  # noqa: E501
        :type host: str
        :param version: The version of this Status.  # noqa: E501
        :type version: str
        :param checked_out: The checked_out of this Status.  # noqa: E501
        :type checked_out: bool
        :param checked_out_since: The checked_out_since of this Status.  # noqa: E501
        :type checked_out_since: datetime
        :param movements_since_checkout: The movements_since_checkout of this Status.  # noqa: E501
        :type movements_since_checkout: int
        :param position: The position of this Status.  # noqa: E501
        :type position: Point
        """
        self.swagger_types = {
            'host': str,
            'version': str,
            'checked_out': bool,
            'checked_out_since': datetime,
            'movements_since_checkout': int,
            'position': Point
        }

        self.attribute_map = {
            'host': 'host',
            'version': 'version',
            'checked_out': 'checkedOut',
            'checked_out_since': 'checkedOutSince',
            'movements_since_checkout': 'movementsSinceCheckout',
            'position': 'position'
        }

        self._host = host
        self._version = version
        self._checked_out = checked_out
        self._checked_out_since = checked_out_since
        self._movements_since_checkout = movements_since_checkout
        self._position = position

    @classmethod
    def from_dict(cls, dikt) -> 'Status':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Status of this Status.  # noqa: E501
        :rtype: Status
        """
        return util.deserialize_model(dikt, cls)

    @property
    def host(self) -> str:
        """Gets the host of this Status.


        :return: The host of this Status.
        :rtype: str
        """
        return self._host

    @host.setter
    def host(self, host: str):
        """Sets the host of this Status.


        :param host: The host of this Status.
        :type host: str
        """

        self._host = host

    @property
    def version(self) -> str:
        """Gets the version of this Status.


        :return: The version of this Status.
        :rtype: str
        """
        return self._version

    @version.setter
    def version(self, version: str):
        """Sets the version of this Status.


        :param version: The version of this Status.
        :type version: str
        """

        self._version = version

    @property
    def checked_out(self) -> bool:
        """Gets the checked_out of this Status.


        :return: The checked_out of this Status.
        :rtype: bool
        """
        return self._checked_out

    @checked_out.setter
    def checked_out(self, checked_out: bool):
        """Sets the checked_out of this Status.


        :param checked_out: The checked_out of this Status.
        :type checked_out: bool
        """

        self._checked_out = checked_out

    @property
    def checked_out_since(self) -> datetime:
        """Gets the checked_out_since of this Status.


        :return: The checked_out_since of this Status.
        :rtype: datetime
        """
        return self._checked_out_since

    @checked_out_since.setter
    def checked_out_since(self, checked_out_since: datetime):
        """Sets the checked_out_since of this Status.


        :param checked_out_since: The checked_out_since of this Status.
        :type checked_out_since: datetime
        """

        self._checked_out_since = checked_out_since

    @property
    def movements_since_checkout(self) -> int:
        """Gets the movements_since_checkout of this Status.


        :return: The movements_since_checkout of this Status.
        :rtype: int
        """
        return self._movements_since_checkout

    @movements_since_checkout.setter
    def movements_since_checkout(self, movements_since_checkout: int):
        """Sets the movements_since_checkout of this Status.


        :param movements_since_checkout: The movements_since_checkout of this Status.
        :type movements_since_checkout: int
        """

        self._movements_since_checkout = movements_since_checkout

    @property
    def position(self) -> Point:
        """Gets the position of this Status.


        :return: The position of this Status.
        :rtype: Point
        """
        return self._position

    @position.setter
    def position(self, position: Point):
        """Sets the position of this Status.


        :param position: The position of this Status.
        :type position: Point
        """

        self._position = position