# server/swagger/swagger.yaml
---
swagger: "2.0"
info:
  description: "A simple REST API to control a meArm robotic arm."
  version: "1.0.0"
  title: "Avanade meArm REST API"
host: "virtserver.swaggerhub.com"
basePath: "/thor-schueler/Avanade.meArm1/1.0.0"
schemes:
- "https"
paths:
  /arm:
    get:
      tags:
      - "Status"
      description: "Gets the current status of the meArm."
      operationId: "get_arm"
      parameters: []
      responses:
        200:
          description: "Successfully retrieved operational status of the meArm."
          schema:
            $ref: "#/definitions/Status"
      x-swagger-router-controller: "swagger_server.controllers.status_controller"
  /arm/position:
    get:
      tags:
      - "Status"
      description: "Gets hte current position of the meArm"
      operationId: "get_position"
      parameters: []
      responses:
        200:
          description: "Successfully obtained position."
          schema:
            $ref: "#/definitions/Point"
      x-swagger-router-controller: "swagger_server.controllers.status_controller"
  /arm/checkout:
    post:
      tags:
      - "Operation"
      description: "Checks out the arm for a session."
      operationId: "checkout"
      parameters: []
      responses:
        200:
          description: "Successfully checkout out the meArm"
          schema:
            type: "object"
            properties:
              token:
                type: "string"
                format: "uuid"
                example: "0729a580-2240-11e6-9eb5-0002a5d5c51b"
                description: "Token to use for any operation in the session"
        403:
          description: "Cannot checkout because arm is checkout to another user."
          schema:
            $ref: "#/definitions/Status"
      x-swagger-router-controller: "swagger_server.controllers.operation_controller"
  /arm/checkin:
    post:
      tags:
      - "Operation"
      description: "Checks in the arm to complete the session and remove the lock"
      operationId: "checkin"
      parameters:
      - name: "token"
        in: "header"
        description: "Session token. This token should be obtained using /arm/checkout."
        required: true
        type: "string"
        format: "uuid"
      responses:
        200:
          description: "Successfully checked in the meArm"
          schema:
            type: "object"
            properties:
              checkedOut:
                type: "boolean"
                example: false
              sessionDuration:
                type: "integer"
                format: "int64"
                example: 3996
                description: "The session duration in seconds"
              numberOfOperations:
                type: "integer"
                format: "int32"
                example: 504
                description: "The number of arm movements performed during the session"
        403:
          description: "Cannot checkin due to incorrect token. Obtain a token using\
            \ /arm/checkout."
          schema:
            $ref: "#/definitions/Status"
      x-swagger-router-controller: "swagger_server.controllers.operation_controller"
  /arm/operate:
    post:
      tags:
      - "Operation"
      description: "Operates the arm using a list of operations."
      operationId: "operate"
      parameters:
      - name: "token"
        in: "header"
        description: "Session token. This token should be obtained using /arm/checkout."
        required: true
        type: "string"
        format: "uuid"
      - in: "body"
        name: "operations"
        description: "A list of operations to be executed."
        required: true
        schema:
          $ref: "#/definitions/Operations"
      responses:
        200:
          description: "Successfully executed operations."
          schema:
            type: "object"
            properties:
              numberOfMovements:
                type: "integer"
                format: "int32"
              duration:
                type: "integer"
                format: "int64"
              position:
                $ref: "#/definitions/Point"
        403:
          description: "Cannot operate due to incorrect token. Obtain a token using\
            \ /arm/checkout."
          schema:
            $ref: "#/definitions/Status"
        413:
          description: "Too many steps. Reduce the number of opertions to 10 or less"
          schema:
            type: "object"
      x-swagger-router-controller: "swagger_server.controllers.operation_controller"
definitions:
  Status:
    type: "object"
    properties:
      host:
        type: "string"
      version:
        type: "string"
      checkedOut:
        type: "boolean"
      checkedOutSince:
        type: "string"
        format: "date-time"
      movementsSinceCheckout:
        type: "integer"
        format: "int32"
      position:
        $ref: "#/definitions/Point"
    example:
      checkedOut: true
      movementsSinceCheckout: 0
      host: "host"
      position:
        r: 5.63737665663332876420099637471139430999755859375
        lng: 7.061401241503109105224211816675961017608642578125
        x: 6.02745618307040320615897144307382404804229736328125
        y: 1.46581298050294517310021547018550336360931396484375
        z: 5.962133916683182377482808078639209270477294921875
        lat: 2.3021358869347654518833223846741020679473876953125
      checkedOutSince: "2000-01-23T04:56:07.000+00:00"
      version: "version"
  Point:
    type: "object"
    properties:
      x:
        type: "number"
        format: "double"
      y:
        type: "number"
        format: "double"
      z:
        type: "number"
        format: "double"
      r:
        type: "number"
        format: "double"
      lat:
        type: "number"
        format: "double"
      lng:
        type: "number"
        format: "double"
    example:
      r: 5.63737665663332876420099637471139430999755859375
      lng: 7.061401241503109105224211816675961017608642578125
      x: 6.02745618307040320615897144307382404804229736328125
      y: 1.46581298050294517310021547018550336360931396484375
      z: 5.962133916683182377482808078639209270477294921875
      lat: 2.3021358869347654518833223846741020679473876953125
  Operation:
    type: "object"
    properties:
      type:
        type: "string"
        enum:
        - "moveTo"
        - "grab"
        - "release"
      target:
        $ref: "#/definitions/Point"
  Operations:
    type: "array"
    items:
      $ref: "#/definitions/Operation"
  inline_response_200:
    properties:
      token:
        type: "string"
        format: "uuid"
        example: "0729a580-2240-11e6-9eb5-0002a5d5c51b"
        description: "Token to use for any operation in the session"
    example:
      token: "0729a580-2240-11e6-9eb5-0002a5d5c51b"
  inline_response_200_1:
    properties:
      checkedOut:
        type: "boolean"
        example: false
      sessionDuration:
        type: "integer"
        format: "int64"
        example: 3996
        description: "The session duration in seconds"
      numberOfOperations:
        type: "integer"
        format: "int32"
        example: 504
        description: "The number of arm movements performed during the session"
    example:
      checkedOut: false
      numberOfOperations: 504
      sessionDuration: 3996
  inline_response_200_2:
    properties:
      numberOfMovements:
        type: "integer"
        format: "int32"
      duration:
        type: "integer"
        format: "int64"
      position:
        $ref: "#/definitions/Point"
    example:
      duration: 6
      numberOfMovements: 0
      position:
        r: 5.63737665663332876420099637471139430999755859375
        lng: 7.061401241503109105224211816675961017608642578125
        x: 6.02745618307040320615897144307382404804229736328125
        y: 1.46581298050294517310021547018550336360931396484375
        z: 5.962133916683182377482808078639209270477294921875
        lat: 2.3021358869347654518833223846741020679473876953125